* chapter7 리액티브 마이크로서비스 개발

+++ microservices/review-service/src/main/java/com/jipps/hombyme/review/services/ReviewServiceImpl.java
@@ -6,14 +6,20 @@ import com.jipps.hombyme.review.persistence.ReviewEntity;
 import com.jipps.hombyme.review.persistence.ReviewRepository;
 import com.jipps.hombyme.util.exceptions.InvalidInputException;
 import com.jipps.hombyme.util.http.ServiceUtil;
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.Flow;
+import java.util.function.Supplier;
+import java.util.logging.Level;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
@@ -26,8 +32,11 @@ public class ReviewServiceImpl implements ReviewService {
 
     private final ServiceUtil serviceUtil;
 
+    private final Scheduler scheduler;
+
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+        this.scheduler = scheduler;
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
@@ -48,10 +57,16 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public List<Review> getReviews(int productId) {
+    public Flux<Review> getReviews(int productId) {
 
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
+        LOG.info("Will get reviews for product with id={}", productId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, Level.FINE);
+    }
+
+    protected List<Review> getByProductId(int productId) {
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
         list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
@@ -63,7 +78,14 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int productId) {
+
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
         LOG.debug("deleteReview: tried to delete reviews for the productId: {}", productId);
         repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
